--- lib/json/detalhes_videos_json.dart
@@ -1,38 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'videos_json.dart';
-
-List detalhesVideosJson = [
-  {
-    'img': 'assets/images/banner_2.png',
-    'titulo': 'Peaky Blinders: Sangue, Apostas e Navalhas',
-    'descricao':
-        'Uma notória gangue da Inglaterra de 1919 é liderada pelo cruel Tommy Shelby, um criminoso disposto a subir na vida a qualquer preço.',
-    'relevancia': '98% Relevante',
-    'ano': '2019',
-    'classificacao': '18',
-    'qtdeTemporada': '5 Temporadas',
-    'videoUrl': 'assets/videos/video_1.mp4',
-    'cor': Colors.yellowAccent,
-    'tipo': 'Trailer',
-    'corClassificacao': Colors.grey,
-    'elenco': 'Cillian Murphy, Sam Neill, Helen McCrory... mais',
-    'episodios': listaVideos[0]
-  },
-  {
-    'img': 'assets/images/banner_1.png',
-    'titulo': 'Lucifer',
-    'descricao':
-        'Entediado no inferno, Lucifer se muda para LA, onde vira gerente de boate e, de quebra, ajuda no combate ao crime. Baseada nos quadrinhos da DC.',
-    'relevancia': '',
-    'ano': '2020',
-    'classificacao': '16',
-    'qtdeTemporada': '1 Temporada',
-    'videoUrl': 'assets/videos/video_2.mp4',
-    'cor': Colors.red,
-    'tipo': 'Prévia',
-    'corClassificacao': Colors.red,
-    'elenco': 'Tom Ellis, Lauren German, D.B. Woodside... mais',
-    'episodios': listaVideos[1]
-  }
-];
