--- lib/json/em_breve_json.dart
@@ -1,59 +1,0 @@
-import 'package:flutter/material.dart';
-
-const List emBreveJson = [
-  {
-    'img': 'assets/images/banner_3.png',
-    'titulo_img': 'assets/images/titulo_2.png',
-    'titulo': 'Castlevania',
-    'descricao':
-        'Um caçador de vampiros luta para salvar uma cidade sitiada por um exército de criaturas controladas pelo próprio Drácula. Inspirado no clássico videogame.',
-    'videoUrl': 'assets/videos/video_3.mp4',
-    'cor': Colors.red,
-    'categorias': [
-      'Sangrento',
-      'Violentos',
-      'Assustadores',
-      'Sombrios',
-      'Animes'
-    ],
-    'data': 'Temporada 4 em 13 de maio',
-  },
-  {
-    'img': 'assets/images/banner_5.png',
-    'titulo_img': 'assets/images/titulo_5.webp',
-    'titulo': 'Ragnarok',
-    'descricao':
-        'Em uma cidade norueguesa envenenada pela poluição e abalada por geleiras derretendo, o fim dos tempos parece real. Só uma lenda poderá combater um mal antigo.',
-    'videoUrl': '',
-    'cor': Colors.grey,
-    'categorias': ['Empolgantes', 'Fantasia - séries', 'Mistério', 'Drama'],
-    'data': 'Temporada 2 em 27 de maio',
-  },
-  {
-    'img': 'assets/images/banner_4.png',
-    'titulo_img': 'assets/images/titulo_4.webp',
-    'titulo': 'O Legado de Júpiter',
-    'descricao':
-        'A primeira geração de super-heróis passa o bastão para os filhos, mas as tensões aumentam — e as regras antigas não se aplicam mais.',
-    'videoUrl': 'assets/videos/video_4.mp4',
-    'cor': Colors.yellowAccent,
-    'categorias': ['Sombrios', 'Empolgantes', 'Super-herói', 'Ação', 'Série'],
-    'data': 'O volume 1 estreia sexta-feira',
-  },
-  {
-    'img': 'assets/images/banner_6.png',
-    'titulo_img': 'assets/images/titulo_6.webp',
-    'titulo': 'Oxigênio',
-    'descricao':
-        'Após acordar e perceber que está presa em uma câmara criogênica, Liz precisa lembrar quem ela é antes que o oxigênio acabe.',
-    'videoUrl': '',
-    'cor': Colors.grey,
-    'categorias': [
-      'Psicológico',
-      'Suspense no ar',
-      'Sci-fi e drama',
-      'Francês'
-    ],
-    'data': 'Disponivel em 12 de maio',
-  },
-];
