--- lib/json/home_json.dart
@@ -1,62 +1,0 @@
-import 'detalhes_videos_json.dart';
-
-List minhaLista = [
-  {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_3.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[0]},
-];
-
-Map minhaLista2 = {
-  'id1': {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[0]},
-  'id2': {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[1]},
-  // ...
-};
-
-List popularesLista = [
-  {'img': 'assets/images/img_3.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_3.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[0]},
-];
-
-List emAltaLista = [
-  {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_3.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_1.webp', 'detalhes': detalhesVideosJson[0]},
-  {'img': 'assets/images/img_2.webp', 'detalhes': detalhesVideosJson[1]},
-  {'img': 'assets/images/img_3.webp', 'detalhes': detalhesVideosJson[0]},
-];
-
-// List originalLista = [
-//   {'img': 'assets/images/img_16.png', 'detalhes': detalhesVideosJson[0]},
-//   {'img': 'assets/images/img_17.png', 'detalhes': detalhesVideosJson[1]},
-//   {'img': 'assets/images/img_18.png', 'detalhes': detalhesVideosJson[0]},
-//   {'img': 'assets/images/img_19.png', 'detalhes': detalhesVideosJson[1]},
-//   {'img': 'assets/images/img_20.png', 'detalhes': detalhesVideosJson[0]},
-// ];
-
-// List animesLista = [
-//   {'img': 'assets/images/img_21.png', 'detalhes': detalhesVideosJson[0]},
-//   {'img': 'assets/images/img_22.png', 'detalhes': detalhesVideosJson[1]},
-//   {'img': 'assets/images/img_23.png', 'detalhes': detalhesVideosJson[0]},
-//   {'img': 'assets/images/img_24.png', 'detalhes': detalhesVideosJson[1]},
-//   {'img': 'assets/images/img_25.png', 'detalhes': detalhesVideosJson[0]},
-// ];
-
-List semelhantesLista = [
-  {'img': 'assets/images/img_2.png'},
-  {'img': 'assets/images/img_3.png'},
-  {'img': 'assets/images/img_4.png'},
-  {'img': 'assets/images/img_5.png'},
-  {'img': 'assets/images/img_6.png'},
-  {'img': 'assets/images/img_7.png'},
-  {'img': 'assets/images/img_8.png'},
-  {'img': 'assets/images/img_9.png'},
-  {'img': 'assets/images/img_10.png'},
-  {'img': 'assets/images/img_11.png'},
-  {'img': 'assets/images/img_12.png'},
-  {'img': 'assets/images/img_13.png'},
-];
