--- lib/json/pesquisar_json.dart
@@ -1,54 +1,0 @@
-import 'detalhes_videos_json.dart';
-
-List pesquisarJson = [
-  {
-    'img': 'assets/images/pesquisar_1.jpg',
-    'titulo': 'The Umbrella Academy',
-    'detalhes': detalhesVideosJson[0]
-  },
-  {
-    'img': 'assets/images/pesquisar_2.jpg',
-    'titulo': 'Orphan Black',
-    'detalhes': detalhesVideosJson[1]
-  },
-  {
-    'img': 'assets/images/pesquisar_3.webp',
-    'titulo': 'One Piece',
-    'detalhes': detalhesVideosJson[0]
-  },
-  {
-    'img': 'assets/images/pesquisar_4.webp',
-    'titulo': 'The Walking Dead',
-    'detalhes': detalhesVideosJson[1]
-  },
-  {
-    'img': 'assets/images/pesquisar_5.webp',
-    'titulo': 'Fate Zero',
-    'detalhes': detalhesVideosJson[0]
-  },
-  {
-    'img': 'assets/images/pesquisar_6.jpg',
-    'titulo': 'Sombra e Ossos',
-    'detalhes': detalhesVideosJson[1]
-  },
-  {
-    'img': 'assets/images/pesquisar_7.webp',
-    'titulo': 'Naruto',
-    'detalhes': detalhesVideosJson[0]
-  },
-  {
-    'img': 'assets/images/pesquisar_8.jpg',
-    'titulo': 'Amor e Monstros',
-    'detalhes': detalhesVideosJson[1]
-  },
-  {
-    'img': 'assets/images/pesquisar_9.jpg',
-    'titulo': 'Baki o Campeão',
-    'detalhes': detalhesVideosJson[0]
-  },
-  {
-    'img': 'assets/images/pesquisar_10.jpg',
-    'titulo': 'Ragnarok',
-    'detalhes': detalhesVideosJson[1]
-  }
-];
