--- lib/json/videos_json.dart
@@ -1,99 +1,0 @@
-const List listaMenu = ['EPISÓDIOS', 'TRAILERS E MAIS', 'OPÇÕES SEMELHANTES'];
-
-const List listaVideos = [
-  [
-    {
-      'img': 'assets/images/detalhe_8.webp',
-      'titulo': '1. Episódio 1',
-      'descricao':
-          'O ambicioso gângster Thomas Shelby percebe que um carregamento de armas desaparecido pode ser sua grande chance de subir na vida.',
-      'duracao': '57m'
-    },
-    {
-      'img': 'assets/images/detalhe_9.webp',
-      'titulo': '2. Episódio 2',
-      'descricao':
-          'Thomas provoca um criminoso local ao trapacear em uma corrida de cavalos e entra em pé de guerra com uma família de ciganos. Campbell organiza um ataque surpresa.',
-      'duracao': '58m'
-    },
-    {
-      'img': 'assets/images/detalhe_10.webp',
-      'titulo': '3. Episódio 3',
-      'descricao':
-          'Thomas pretende se aproximar de Billy Kimber nas corridas de Cheltenham e considera uma oferta de simpatizantes do IRA para comprar suas armas.',
-      'duracao': '55m'
-    },
-    {
-      'img': 'assets/images/detalhe_11.webp',
-      'titulo': '4. Episódio 4',
-      'descricao':
-          'Thomas está cada vez mais desconfiado da noiva de seu irmão, e Campbell pressiona a gangue a entregar as armas.',
-      'duracao': '58m'
-    },
-    {
-      'img': 'assets/images/detalhe_12.webp',
-      'titulo': '5. Episódio 5',
-      'descricao':
-          'Thomas lida com um líder do IRA querendo se vingar da morte de seu primo, e Grace precisa decidir se é leal a Campbell ou Thomas.',
-      'duracao': '56m'
-    },
-    {
-      'img': 'assets/images/detalhe_13.webp',
-      'titulo': '6. Episódio 6',
-      'descricao':
-          'Thomas se prepara para tirar Billy Kimber do poder, mas segredos de família e a obsessão de Campbell podem atrapalhar seus planos.',
-      'duracao': '54m'
-    },
-  ],
-  [
-    {
-      'img': 'assets/images/detalhe_1.webp',
-      'titulo': '1. Piloto',
-      'descricao':
-          'Depois que um velho conhecido tem uma morte violenta, o charmoso Lucifer jura vingança contra os culpados e consegue uma aliada inesperada.',
-      'duracao': '44m'
-    },
-    {
-      'img': 'assets/images/detalhe_2.webp',
-      'titulo': '2. Fica, Lucifer. Bom Diabo.',
-      'descricao':
-          'Apesar da vontade de voltar para o inferno, Lucifer decide investigar de que forma os paparazzi contribuíram para a morte de um jovem.',
-      'duracao': '43m'
-    },
-    {
-      'img': 'assets/images/detalhe_3.webp',
-      'titulo': '3. O quase Príncipe das Trevas',
-      'descricao':
-          'Um quarterback profissional prestes a virar a bola da vez se mete na maior encrenca e pede para Lucifer ajudá-lo a provar sua inocência.',
-      'duracao': '43m'
-    },
-    {
-      'img': 'assets/images/detalhe_4.webp',
-      'titulo': '4. Novidades masculinas',
-      'descricao':
-          'Lucifer tenta seduzir Chloe, mas o que dá certo com as outras não funciona com ela. Amenadiel confessa a Maze que está preocupado com Lucifer.',
-      'duracao': '42m'
-    },
-    {
-      'img': 'assets/images/detalhe_5.webp',
-      'titulo': '5. Sapatos',
-      'descricao':
-          'Lucifer descobre que nem tudo no mundo da moda é belo. Após conseguir novas informações, Amenadiel bola um plano para levar Lucifer para casa.',
-      'duracao': '43m'
-    },
-    {
-      'img': 'assets/images/detalhe_6.webp',
-      'titulo': '6. Filho favorito',
-      'descricao':
-          'Quando algo pessoal e valioso para Lucifer é roubado, ele se envolve na investigação. Enquanto isso, a Dra. Martin toma uma importante decisão.',
-      'duracao': '43m'
-    },
-    {
-      'img': 'assets/images/detalhe_7.webp',
-      'titulo': '7. Parceiro',
-      'descricao':
-          'Lucifer continua procurando o container roubado e recebe ajuda de quem menos esperava. Chloe pede para Dan ajudá-la no caso Palmetto.',
-      'duracao': '43m'
-    },
-  ]
-];
