+++ lib/ok/menus/menu.dart
@@ -2,20 +2,15 @@ import 'package:flutter_view_controller/flutter_view_controller.dart';
 import '../../components/custon_button/custon_button.dart';
 import 'package:flutter/material.dart';
 
-import 'menu_services.dart';
-import 'menu_stores.dart';
+import 'menu_home.dart';
 import 'menu_test.dart';
 
 class MenuPetController extends Controller {
   @override
   onInit() {}
 
-  goMenuStores() async {
-    Navigator.pushNamed(context, (MenuStoresView).toString());
-  }
-
-  goMenuServices() async {
-    Navigator.pushNamed(context, (MenuServicesView).toString());
+  goMenuHome() async {
+    Navigator.pushNamed(context, (MenuHomeView).toString());
   }
 
   goMenuTest() async {
@@ -54,14 +49,7 @@ class MenuPetView extends ViewOf<MenuPetController> {
             children: <Widget>[
               SizedBox(height: size.height(14)),
               customButton(
-                  text: 'Stores',
-                  onPressed: controller.goMenuStores,
-                  size: size),
-              SizedBox(height: size.height(1)),
-              customButton(
-                  text: 'Services',
-                  onPressed: controller.goMenuServices,
-                  size: size),
+                  text: 'Home', onPressed: controller.goMenuHome, size: size),
               SizedBox(height: size.height(1)),
               customButton(
                   text: 'Testes', onPressed: controller.goMenuTest, size: size),
