+++ lib/ok/menus/menu_test.dart
@@ -2,21 +2,14 @@ import 'package:flutter_view_controller/flutter_view_controller.dart';
 import 'package:flutter/material.dart';
 
 import '../../components/custon_button/custon_button.dart';
-import '../tests/container.dart';
-import '../tests/container2.dart';
-import '../home/home.dart';
-import '../home/button_navigation.dart';
-import '../tests/container3.dart';
-import '../tests/container4.dart';
-import '../tests/listview_image.dart';
+import '../../screens/tests/container.dart';
+import '../../screens/tests/container2.dart';
+import '../../screens/tests/container3.dart';
+import '../../screens/tests/container4.dart';
 
 class MenuTestController extends Controller {
   @override
   onInit() {}
-
-  goHomeView() async {
-    Navigator.pushNamed(context, (HomeView).toString());
-  }
 
   openContainer() async {
     Navigator.push(
@@ -46,20 +39,6 @@ class MenuTestController extends Controller {
     );
   }
 
-  openListviewImage() async {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => ListViewImage()),
-    );
-  }
-
-  openbottomBar() async {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => ScreenHome()),
-    );
-  }
-
   @override
   onClose() {}
 }
@@ -82,16 +61,6 @@ class MenuTestView extends ViewOf<MenuTestController> {
             children: <Widget>[
               SizedBox(height: size.height(14)),
               customButton(
-                  text: 'Home Solo',
-                  onPressed: controller.goHomeView,
-                  size: size),
-              SizedBox(height: size.height(1)),
-              customButton(
-                  text: 'Page Botton',
-                  onPressed: controller.openbottomBar,
-                  size: size),
-              SizedBox(height: size.height(1)),
-              customButton(
                   text: 'Container Test',
                   onPressed: controller.openContainer,
                   size: size),
@@ -109,11 +78,6 @@ class MenuTestView extends ViewOf<MenuTestController> {
               customButton(
                   text: 'Container Test 4',
                   onPressed: controller.openContainer4,
-                  size: size),
-              SizedBox(height: size.height(1)),
-              customButton(
-                  text: 'Listview Image',
-                  onPressed: controller.openListviewImage,
                   size: size),
               SizedBox(height: size.height(1)),
             ],
